demo: burgers -> burgers1; mesh_seq -> burgers
