+++ makefile
@@ -23,7 +23,9 @@ demo:
 	@echo "Running all demos..."
 	@cd demos && python3 time_partition.py
 	@cd demos && python3 burgers.py
+	@cd demos && python3 burgers1.py
 	@cd demos && python3 burgers2.py
+	@cd demos && python3 burgers3.py
 	@echo "Done."
 
 doc:
